Merged PR 588612: Reconcile skip through db
Instead of looking through all the content we have in the db, we skip to the lesser of the last processed hash between adds and removes.

If only one of the last processed hashes has a value we skip to that value, since the other operation would have went through all the content.