Merged PR 587441: Use more aggressive bandwidth settings for the first half of all copy attempts
This PR addresses the following work item #1791653
Basically, this PR starts limiting the usages of more aggressive bandwidth configuration only for the first half of the attempts.

For instance, if the hash has > 32 locations its possible to use aggressive bandwidth configuration for all the attempts. With this change we'll be using only for the first 16 attempts.

Related work items: #1791653