Merged PR 589902: Set TotalOrderSeek option during database enumeration
This PR sets `SetTotalOrderSeek` value on `ReadOptions` to true during database enumeration to avoid reading stale data from the database.

When the prefix extractor is used then `SetTotlaOrderSeek` must be set to true to avoid reading stale records from the database.

This is a very important change, because the old was causing the issues during reconciliation (and, most likely, during GC as well), when the result of the database enumeration was out of sync with the actual state of the world.

The following doc sheds some light on this topic: https://github.com/facebook/rocksdb/wiki/Prefix-Seek#general-prefix-seek-api