Merged PR 590890: Add an implicit DScript resolver that owns all in-box SDKs
This PR adds an implicit DScript resolver after all user configured ones that owns all in-box SDKs that are deployed together with bxl binaries. These are usually Sdk.Transformers, Sdk.Drop, etc.
Observe that if any of the in-box ones are explicitly provided by the user, those will take precedence as this is the way the workspace resolves name clashes.
The main goal of this feature is simplify the configuration of small repos (e.g. JS like ones) so basic SDKs can be used without explicitly mentioning them.