Merged PR 591788: Disconnect RemoteWorker before disposing in MasterService
If the RemoteWorker send thread is started (after attaching to master) and the build is terminated without executing (e.g. if no pips match the filter), the thread is alive when we call `Dispose()`, which triggers a contract assertion failure (send thread shouldn't be alive).

This ensures the worker is stopped (i.e. the thread is closed) in the MasterService before being disposed.

Related work items: #1801675