Merged PR 591894: Use service name in LifetimeTracker to avoid concurrency issues when Launcher...
This PR fixes the issue that is happening with `LifetimeTracker` when the launcher is used.

When the launcher is used, then two `LifetimeTracker` instances are created (one by casaas process and another one by cache service process) and each of them were using the same file for tracking the lifetime of the service.

This was causing a lot of issues, like concurrent writes to the same file, failure to move the files around and contract violation trying to write the file on disk.

Related work items: #1805962