Merged PR 592593: Handle authentication against ADO for the download resolver
Add to the download resolver basic authentication to download packages from ADO feeds. This includes:
* Authenticating via a credential provider, which covers the CB case
* Authenticating via IWA, which covers domain joined accounts (mostly for local builds)

The auth token is only used against ADO URIs over https.

Tried to add tests but it is pretty complicated to get reasonable IWA/credential provider mocks to make it work. And at the same time, this feature looks simple enough to not deserve that much work on the test infra.

Related work items: #1807181