Merged PR 590571: Add stable verification for chunk deduped files
## Context
Users are running into hash verification failures when downloading drops due to (required) behavioral changes we've made to the internal logic of our chunkers. The file, when chunked using the new version of the chunker, results in a different chunk configuration, which causes the overall file content verification to fail.

## Solution
Based on Brian's idea, this PR introduces a new verifier that verifies the file's hash by chunking the file according to it's original chunk configuration. The same chunks of equal offset, sizes and hash are produced, and thus the overall hash produced is the same. This allows for the overall file content to be verified correctly as a result.

## Test coverage
- Updated existing test cases for hashes (`DedupContentHasherTest` which verifies hash produced of a large variety of different datasets)
- Add new unit test coverage
- Based on Kunal's feedback, manually validated solution works against a real-live production drop in ICM https://portal.microsofticm.com/imp/v3/incidents/details/214970854/home.
- Performance testing against a single production file of 5GB with 1MB chunks indicates the verification is 2x faster than using the chunker algorithm to verify the file content

See related uptake !590195

Related work items: #1797540