Merged PR 592835: Fix MsBuildGraphBuilder conflictive assembly loading of System.Collections.Immutable
The MsBuildGraphBuilder tool was loading `System.Collections.Immutable.dll` from the tool deployment folder. In newer versions of MSBuild a different version of the library is being referenced: loading this newer MSBuild and the old DLL caused a runtime error, (e.g. our example MsBuild project failed to build with the latest MsBuild with a `MissingMethodException`).

Related work items: #1807610