Merged PR 592613: Fix FileNotFound errors happening from BlobCentralStorage during checkpoint restore
This PR fixes an issue when checkpoint restore may fail with FileNotFound exception, because the folder where the files are copied from blob central storage is missing.

The error is happening only when the first sst file comes from the blob central storage. In this case `_checkpointStagingDirectory / "chkpt"` is not created and the call to `BlobCentralStorage.TryGetFileCoreAsync` will fail with 'FileNotFoundException' (even though the file exists, but the part of the directory is missing).

Related work items: #1806838