Merged PR 593700: Update IcM incident TTL in the monitor on every hit
Currently we just returned and did nothing. Now, we bump the TTL on every hit.

This opens us up to a bad scenario:
1. Redis falls into bad state and we emit an IcM.
2. We fix the Redis instance and mark the incident as resolved.
3. The same Redis instance fails again within a day of the last time the monitor checked it.
4. No incident is emmited even though it is a new incident.

However, with the new logic it should be safe to have a much smaller TTL for avoiding repeats. This is why I modified Redis autoscale issues TTL from 1 day to 1 hour.