Merged PR 593737: Retry attachment to workers until WorkerAttachTimeout is reached
Retry attachment to workers from the orchestrator until `WorkerAttachTimeout` is reached.

This addresses #1782253 which is happening because the orchestrator is trying to attach to workers that are still finishing a previous build session (so the RPC goes through but the call fails because the build ids don't match). After that failure the orchestrator doesn't try again and, even though the workers become available later, they are never attached to.

Related work items: #1782253