Merged PR 595251: Cancel remote worker's send thread on connection timeout
Currently, when `Exit`ing the worker due to connection timeout we end up joining the worker's send thread, which in turn is using said connection to send the results. If those RPCs timeout/fail (which is to be expected if we're in the middle of a connection timeout callback), then that thread will also call `Exit`. Because this is all fire-and-forget there isn't any catastrophic errors / circularity, but there is a race in the error message setting (and also unnecessary context switching, waiting for those RPCs to fail, etc.).

This PR adds a cancellation token which is triggered by the connection timeout callback, trying to terminate the send thread as early as possible and also avoiding multiple calls to `Exit`.

Related work items: #1807760