Merged PR 596292: Untrack copy of local shim when pip executes in VM
It's an .exe that can be injected using Image File Execution Options to intercept processes such as LSBuild.exe that don't work well when executed from the network mapped D: drive. It will copy the original .exe and its entire folder into the VM and re-execute the original command-line from there.  It is meant to be abstracted away from the customer's build, it's used as a compatibility measure and therefore the customer's build shouldn't know about it.

Related work items: #1814190