Merged PR 595517: Add existence assertions under opaque directories
This feature allows to assert the existence of an output under an opaque directory and consume it as a regular file artifact downstream. The existence check is deferred to the execution phase. E.g.

```
const outputDir : StaticDirectory = ScheduleSomeProcessPip().getOutputDirectory(d`Out`);
const outputFile : File = outputDir.assertExistence(a`foo.txt`);
const consumer = ConsumeOutputFile(outputFile);
```
Main use case comes when doing package install in JS repos: npm install is executed to get some tool (e.g. rush, yarn) and then the tool needs to be executed. This implies getting a ToolDefinition pointing to a file that was produced in the opaque