Merged PR 598135: [SandboxConnectionLinuxDetours] Make it resilient to filesystem crashes
SandboxConnectionLinuxDetours uses a FIFO to send access reports from the sandbox back to the bxl process.

Calling 'read' on a FIFO blocks until all writers have been closed.

To make sure reading access reports terminates, SandboxConnectionLinuxDetours opens a dedicated write handle for the same FIFO; once it closes that handle, 'read' returns 0 which signals that there are no more reports to be received.

This works fine except in some extreme cases. One such extreme case is when the underlying filesystems crashes or shuts down completely (which is possible, especially if that's a custom-implemented filesystem running in user space). When that happens, not all write handles to said FIFO are necessarily going to be closed, so our 'read' call may remain stuck forever.

To work around this case, this PR makes it optional to wait for the "read thread" to complete.