Merged PR 598956: Optionally use low allocation NLog layout
This PR introduces an optional low-allocation NLog layout to reduce the allocations that happen when writing every message to the logs.

Currently, due to a weird NLog layering its hard to pass a `StringBuilder` to `Layout.Render` method. To do that a new class that derives from `SimpleRenderer` is introduced and its used if configured.

Here is the result of the benchmark:

```
|   Method |       Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------- |-----------:|---------:|---------:|-------:|------:|------:|----------:|
| Original | 1,006.7 ns | 30.18 ns | 84.14 ns | 0.0572 |     - |     - |     244 B |
|    Fixed |   913.1 ns | 17.96 ns | 37.88 ns | 0.0257 |     - |     - |     108 B |
```

From profiling session (this PR should kills like 10% of all transient allocations - we can avoid 185Mb out of 1.75Gb):
![image.png](https://dev.azure.com/mseng/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/598956/attachments/image.png)

Related work items: #1819095