Merged PR 599075: Update NLog to 4.7.7 and use built-in prop to avoid excessive allocations
It seems that the hack I've added to avoid allocations is not needed.

One of the NLog maintainers opened a bug: https://github.com/microsoft/BuildXL/issues/1283 with the explanation of the right solution and this PR uses the proposed solution instead of relying on a custom target I've created previously.