Merged PR 597926: Refactor notification sending to unify messages in a single queue
Creates a `NotificationManager` as a point of aggregation for all events. The three different streams of messages we had before, which were directly sending the Notify RPC, instead report the messages to this manager which unifies the payloads.

This means that all execution log and forwarded events reach the orchestrator before or with the result of the pip that produced them. The binary execution log data is now sent along with the pip results (in fairly small chunks) instead of independent 1MB blobs.

Pending:
- Avoid copying bytes from buffer to buffer for the execution log data. Prior to this refactor, the execution log target's thread blocked while sending the RPC so it could safely use a view of the underlying buffer (via `ArraySegment`) while this was happening. Now we send these data in a different thread so it's being copied to avoid missing data (because we want to clear the buffer after sending the data).

Related work items: #1816726