Merged PR 600595: Increase API Server concurrency
API server uses ActionBlock to control how many clients are connected. Each of the daemons creates a client when it starts; since daemons are service pips, the process (and the client) lives for quite some time. If we have multiple daemons, each daemon takes a slot in the action block. Once we reach the concurrency limit, we new clients (and requests that they must serve) are waiting in the queue blocking pips and essentially hanging a build.

This PR increases the max number of clients that can be connected at the same time, and it also makes "maxRequestPerClient" a separate setting. It's mostly done to cover a corner case when there are many daemons hammering the api server at the same time. If we to use the same setting for both of this settings, we'll need to use n^2 threads just to process the requests.

In a future PR, we'll need to add special handling for a case when we have more clients than the concurrency limit hardcoded in the scheduler.

Related work items: #1812805