Merged PR 600605: Dispose event listeners only after finishing the worker's send thread
This removes a couple of race conditions around the flushing of the execution log target. Disposal forced a "manual" (i.e. not queued in the action queue of the target) flush of the log target (on Close()), which could conflict with the flushing done by the send thread (which is the only one that should be flushing it while it is alive). Moving it down after joining the send thread removes the race, without affecting correctness.

Related work items: #1820523