Merged PR 601579: [SandboxConnectionLinuxDetours] Cancel continuations as soon as the object is disposed
Before this change, after a pip completes we'd fire and forget a task that waits for a minute and then makes sure that we've received all access reports for that pip.  Most of the time, the reports are received right after the pip completes so we end up unnecessarily holding a reference to that pip (and all the transitively reachable objects) for a whole minute.  In extreme cases, that can cause non-negligible memory pressure.

This PR adds a cancellation token source per pip that gets triggered as soon as the pip completes, immediately cancelling the wait task mentioned before.