Merged PR 601774: Don't send any more events to the orchestrator when pip results queue is marked as complete
The error/warning events are being forwarded alongside pip results. When the orchestrator calls `Exit` on a remote worker, the _pip-results_ queue is marked as complete and drained, together with any events that were already in the _outgoing-event_ queue. After this, the outgoing-event queue is marked as complete and the cancellation token for RPC calls is signalled.

This PR makes it so that any events processed after the pip-result queue is completed and drained are effectively ignored.

There are situations in which events may arrive to this queue after this pip-result queue is drained. For instance, events may be happening with the execution of MaterializeOutput steps in a "fire-forget materialize output" build. Because the orchestrator calls `Exit` while execution is still ongoing in the worker, these events may get to the event listener after sending our final pip result from the pip-result queue but before signaling the token.

Because we are in the middle of the shutdown process, we can assume that these events are safe to ignore. The additional assumption is that any events that the orchestrator does care about (e.g. in the previous situation, any events regarding previous pip steps before MaterializeOutputs) will be in the outgoing-events queue before the corresponding pip result gets to the pip-result queue, so that we don't miss any events.