Merged PR 605369: Make the reparse point cache case insensitive
During Rush package install there are cases where the same path is exercised with different casing. E.g.:

1. Probe path/to/foo/file.txt (absent)
2. Create path/to/FOO as a reparse point
3. Probe path/to/foo/file.txt (now present)

Access 1 populates the cache with path/to/foo/file.txt as not needing resolution. Access 2 tries to invalidate that path, but casing doesn't match and therefore the path remains in the cache. Access 3 should get a resolved path, but the cache is still populated with a 'do not resolve' result for it.

This PR makes the reparse point cache case insensitive by normalizing casing. It also takes care of trailing slashes, which have a similar effect to casing.