Merged PR 606905: Proactively copy checkpoint files on checkpoint creation.
This is to reduce copy load on the master.

With the current implementation, we will not be reusing the logic in ReadOnlyDistributedContentSession. This is just to be able to get an idea of the effectiveness of this feature before doing more disruptive refactors to be able to share the logic.

Related work items: #1832898