Merged PR 594831: Use hierarchical trace ids if configured
This PR changes (if the flag is one) the way the nested tracing ids are created.

If the flag is one then instead of just creating a guid every time, the nested id is constructed as 'parent_id'.currentId that allows tracking all the call tree by using the following queries: `where CorrelationId has 'parent_correlation_id'`.

I'm currently running those bits in DM_S2 to see how useful this feature is.