Merged PR 607317: [LinuxSandbox] Implement a file descriptor table
To get an absolute path for a file descriptor, we used to call `readlink("/proc/self/fd/{fd}")`.  That's perfectly fine and still the best way to do it; the problem is that it is slow.  (for example, it is not uncommon that a C compiler calls `putc` millions of times to produce a build output).

This PR implements a file descriptor table where where a mapping from a file descriptor to an absolute path is cached.

A new entry is computed and inserted into the table whenever a path for a file descriptor is needed but it hasn't already been computed.

Because file descriptors may be reused for different files over time, whenever a file descriptor is closed (i.e., the `close` syscall is intercepted) the corresponding entry in the table is invalidated.

Related work items: #1833259