Merged PR 608900: Cancel running processes if there are no suspended ones when memory usage is critical
The memory manager uses effective RAM to decide if pips should be suspended / cancelled. When there is a big modified page file, the actual memory usage may be very high but we optimistically don't do anything, expecting some relief soon from paging.

However, if actual RAM usage is >98% and the modified pagefile accounts for more than 50% of the RAM we do start cancelling _suspended_ pips (without taking into account the effective memory usage) because we assume that thrashing is starting to become an issue. If there aren’t any suspended pips to cancel, this basically amounts to doing nothing. Telemetry shows that we can hit this situation and builds can spend a long time in a critical-memory state with no progress.

This PR allows cancellation of running pips in this critical state too.

Related work items: #1832915