Merged PR 605118: Decouple WorkerService's "distribution" and "execution" logic and add unit tests for distribution
Split WorkerService into a "front end", which handles initialization/finalization and communication, and a private "back end" which does the work for every pip and communicates with the actual engine. In this way we can unit test the service in different communication scenarios by mocking both the engine and the server.

Related work items: #1824124