Merged PR 609251: Fix PathCombine in Detours
`PathCchCombineEx` apparently has unexplained semantics. Depending on the path and the buffer size, it sometimes preserves `\\?\` NtEscape, but sometimes it removes it.

Prefix replacement from the resulting path combine can truncate the paths when `\\?\` is removed.

Using plain string semantics for path combine is the way to go.

Related work items: #1836779