Merged PR 609190: Delete present directory before materializing a reparse point
When materializing reparse points we don't handle the case where there is a present directory in that path. When this happens the reparse point creation fails. This PR fixes this failure by deleting the contents of the path (including when it is a directory ) before creating the reparse point.

Additionally, the new logic avoids opening some file handles as we already know if the path is present / absent, addressing
 #1823031

Related work items: #1823031, #1834496