Merged PR 609358: Rename "master" to "orchestrator" in distribution code
Rename "master" to "orchestrator" in distribution code. This PR renames everything in the code which is "internal" and safe to replace:

- variable/class names
- comments
- log event ids and method names
- documentation

For the following:
- `DistributedBuildRoles.Master` enums (the names are used to parse arguments)
- program arguments:  `masterCpuMultiplier`, `masterCacheLookupMultiplier`

New arguments `orchestratorCpuMultiplier` and `orchestratorCacheLookupMultiplier`, and `DistributedBuildRoles.Orchestrator` were added so we can transition customer seamlessly and then remove the old one.

Still pending due to the changes being more sensitive are

- protobuf interfaces / gRPC generated classes like `Master.MasterClient`, and names in bond objects
- DScript options (essentially `mustRunOnMaster`)

Related work items: #1745844