Merged PR 609442: Use in-ring machines for file copies
This PR implements an ability to use in-ring machines for copying files.

The idea is that in some cases the location data is not available for some reason, and the builds are failing that could've been successful if the cache service would've considered in-ring machines as the candidates for locations.

This PR implements this logic and traces explicitly if the copy was successful from an in-ring machine.

Related work items: #1822996