Merged PR 610525: Update the calculation of orchestrator's slots in the distributed builds
In the old calculation method (1/n), the concurrency limit was decreasing faster than it should be. It hurts the performance of low-worker distributed builds such as JS-repos because the orchestrator CPU multiplier is 0.5 in 2-worker JS builds. We already overcome this in 3-worker and 5-worker Cosine builds by passing orcheastratorCpuMultiplier. However, we want to fix it by default.

To avoid the steep curve, we changed the calculation method.

Related work items: #1836729