Merged PR 610205: Memory optimizations for ContentHash and ShortHash
This PR reduces the size of `ShortHash` struct from 33 bytes to 12, that cut the memory consumption in many places in the code. One notable place is `VolatileSet<ShortHash>` that is used by casaas master where the memory footprint is down by 100Mb for 500K keys.

This PR also reduces the size of `ContentLocationEventData` that stores the short hashes in the list that reduced the overall memory traffic for casaas master by tens of MB per seconds during heavy load (including hitting small heap instead of the LOH).

This PR also adds allocation-free version for serialization and getting `byte[]` for `ContentHash` and `ShortHash` that cut tens of MB allocations per second.

Related work items: #1838512