Merged PR 610794: Add ToAwaitable extension for CancellationToken
`TaskExensions.WaitForCancellationAsync(this CancellationToken)` was designed to create a task from `CancellationToken` when the token is canceled.
Unfortunately, the original implementation had some issues: if the token is never canceled, a timer created by `Task.Delay` was never disposed. So calling this method many times may overwhelm the timer queue.

Another potential approach for converting `CancellationToken` to `Task` is based on `TaskCompletionSource` when the source is canceled from `token.Register()` method. But this approach is also problematic, because there is no way to "unregister" if the token is never canceled.

To work-around the aforementioned issue, the solution is to return a disposable struct which `Dispose` method will unregister from the token successfully even if the token is never canceled.