Merged PR 611920: [Detours] No path cache when ignoring full symlink resolving
This PR completely disables path cache when ignoring full symlink resolving.

When we don't fully resolve symlinks, the cached resolved paths can be invalidated incorrectly (see unit test). In the test the invalidation of the middle directory symlink does not invalidate all possible paths reaching to the file symlink, and thus the cache retains incorrect resolved path. In this case, `D2\y.txt` is not reported although the symlink `D.lnk` now points to `D2\`, instead of `D1`.

This PR also re-applies the reverted PR !609600.

This PR may address the slow symlink creation reported by customer even when fully symlink resolving is disabled. The slow creation can be attributed to the lock contention in the path cache. But now, it is completely disabled when fully resolve symlink is disabled.

Related work items: #1840718