Merged PR 612960: Use only one BuildCache and local content session per publishing session
Currently we have a pool of `BuildCacheSession`, which results in multiple BuildCache clients. We could have only one of them per publishing session, and that should result in better resource usage.

Same thing for local content sessions. Currently we open a session every time a content hash lists requires us to push content. This seems like overkill to me, and tracing for publishing operations is extremely chatty because of this. Limiting to one content session per publishing session. This may have some unwanted effects, such as pinning content for longer than we usually would, or even opening a session when it otherwise wouldn't be necessary, but I believe this is a lot cleaner and, unless we see that pinning becomes a problem, I don't see this hurting us in a significant way.

When applying these changes, it became clear that it made more sense to move the publishing logic from the store to the session, while making sure that all sessions share a single IO gate.

Note that implementations for `CreatePublisher`, `PublishContentHashListAsync`, and `PushToRemoteAsync` are exactly the same in the session as they were in the store, with the only differences being how we acquire BuildCache and content sessions

Related work items: #1843035