Merged PR 613292: Add support to the SymbolDeamon for JavaScript SourceMap symbols.
Add support to the SymbolDeamon for JavaScript SourceMap symbols.

JavaScript development is getting more and more common at Microsoft.
For native and for managed development we have gotten used to have
centralized management and processing of symbols (.pdb) files.

JavaScript also supports an equivalent of symbols called SourceMap.
For Office engineers it is getter important to have the symbols accessible
so we can debug javascript files. Storing these in the same place as
pdbs makes the most sense from storage, security and all other aspects.

The azure devops backend is completely agnostic about the file contents,
it jsut delegates to the client code to complute a clientid and then processes
it.

For regular pdb's the clientkey is embedded by the compiler in the .pdb, so the
client code that uploads the symbols cracks open the pdb to get the client key
out of it and uses that to upload it.
JavaScript sourcemaps are just JSON files and can be extended with fields starting
with "x_<companyname>_". I'm doing so and have written plugins to the javascript
tools to emit this. As a fallback the code will use the sha256 content hash
of the .js.map file.

The code I'm adding here actually belongs in the Symbol library owned by Azure Devops
but they are in the process of rewriting the tool/library and make it
open source. This work is still about 2 months out for me to add the logic
there. I will port the code there when it is ready so that vanilla azure devops
pipelines will get support for .js.map symbols as well.

I would like to make progress for office builds on this, so that is where this PR
comes in. It allows me to get started and unblock the office devmain builds on cloudbuild
If I did the work to update the code in Azure DevOps I would still have to send a PR
to update BuildXL to use the newer library.

When the ADO has a release with the new bits I will come back to update buildxl,
remove this code and delegate to the ADO logic.

Related work items: #1844922