Merged PR 613396: Improve perf in CopyToWithFullBufferAsync by setting the file length if available
This change sets the file length if available to improve `PutStream` performance.

The change adds the test that should be run manually that shows the performance difference when the file length is set or not set upfront.

```
WriteToTemporaryFiles with no setting lengths: 17163.5408ms
WriteToTemporaryFiles with setting file length: 12151.7383ms
```

Her is the snapshot from the profiler:

Before:
![image (2).png](https://dev.azure.com/mseng/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/613396/attachments/image%20%282%29.png)

After:

![image.png](https://dev.azure.com/mseng/9ed2c125-1cd5-4a17-886b-9d267f3a5fab/_apis/git/repositories/50d331c7-ea65-45eb-833f-0303c6c2387e/pullRequests/613396/attachments/image.png)

The wallclock time is not comparable between the two runs, but the diff shows that `SetLengthCore` was expensive and that its gone when the length is set upfront.

Related work items: #1844115