Merged PR 611665: Add overflow buffers to increase capacity in the string table
In the StringTable we are already using a "one string per slot" buffer (the LargeStringBuffer) to store large strings: the `StringId` associated to this LargeStringBuffer indexes it by slot number instead of by an offset. This PR allows the user of a string table to specify an arbitrary number of "overflow buffers" which work the same way as the LargeStringBuffer: it has a backing `byte[][]` which will hold an individual string in every position.

With this, the main table is reduced in `OverflowBufferCount` rows, but every overflow buffer can hold 2^21 new strings,  with the cost of extra allocation of byte arrays, so the overall capacity is increased. These allocations only happen if we reach full capacity of the 'main' table. This shouldn't be used by almost anyone, but we have already observed large builds which are filling the string table:

#1840418

Related work items: #1840418