Merged PR 614181: [Perf] Always set file length if available when opening files for writes
Instead of setting the file length in various places, this PR introduces and extensions methods `AbsFileSystemExtensions.TryOpenForWrite` and `OpenForWrite` that takes an expected file length as an argument.

I tested the copy speed and got roughly the same performance improvement that I got in the other PR: ~20-30% faster when the file length is set upfront.

This PR is so massive because the new API for opening files is synchronous and I decided to remove all the async open file API altogether (the API is no longer asycnhronous once we removed a semaphore from `PassThroughFileSystem`).

Related work items: #1845340