Merged PR 613409: Close client channel when detecting RelatedActivityId mismatch
Currently when detecting a build id mismatch we only cancel the current call and will keep retrying further communications. This PR makes it so the server signals an "unrecoverable mismatch" when detecting that the caller's RelatedActivityId doesn't match the server's -- in this case the client will close the connection.

When the RelatedActivityIds match but the environments don't the previous behavior is preserved: we want the orchestrator server to keep trying to connect to workers that will become available later and but may be participating in a previous build (e.g. for Office builds with `/fireForgetMaterializeOutput+`, the worker is still in the MetaBuild materializing outputs and the orchestrator already started attachments in the ProductBuild).

 The PR also adds unit tests for this behavior and some improvements in other gRPC tests.

Related work items: #1839858