Merged PR 616167: Optimize searching for reparse points in a path
Optimize searching for reparse points in a path by starting from the first level in the path that the policy result says to enforce full reparse point parsing for.

If a user specifies to enforce reparse point access under d:\office\import\x64, then we shouldn't try to look for reparse points for any path less than that.  To accomplish this the level of the path is kept (d: = L0, d:\office = L1, etc...) and the first level to enable reparse point enforcement is computed when searching for a policy cursor.  During path iteration, don't check for reparse points until the level >= first level to enable reparse point parsing for.