Merged PR 619258: Allow querying for static mounts during JS custom script evaluation
During JS custom script evaluation the mount table is already initialized but not finalized. Relax the assertion that requires the mount table to be finalized before querying to limit the exposed mounts in this case to the ones that are statically defined. This avoids non-deterministic behavior, where module-level mounts may be added before finalization, but allows basic mounts to be exposed for computing JS scripts

Related work items: #1853025, #1853611