Merged PR 616549: Add an orchestrator-side pip timeout period for remotely scheduled pips
Implement a timeout for pip results to come back from workers. If after 5.25 hours we don't hear back from a pip, disconnect the worker. Also adds an integration test to exercise the behavior.

The 5.25 hour default is chosen because our longest pip timeouts are set to 5h for some queues, and this is meant to be a failover that kicks in when we lost the pip result for some reason.