Merged PR 617865: [dropd] Allow customization of drop paths for directory content
Allow customization of drop paths when users add directory artifacts. Essentially, this PR allows dropping files inside of subDir of the original artifact at any location.

Due to the way the validation builds work, I'll have to split the change into two PRs. Once this one is merged and is a part of a deployed LKG, another PR will enable the feature on the front end side.

Related work items: #1845257