Merged PR 620851: GetBulk tracing improvements and filtering out inactive machines for GetBulk(Global)
This PR improves the tracing for `GetBulk` calls in the following way:
1. Prints out if the checkpoint is stale (older than the TTL in Redis)
2. Prints inactive machines filtered out from `GetBulk` result (the list of up to 3 machines and the count)

3. Filters out inactive machines for `GetBulk(Global)` call.
After the discussion with Lance, this change was removed: we do register content into the global store once the machine moves from Inactive to Active state. This means that if we filter the inactive machines for the global store as well, the location data will be unavailable for a longer period of time potentially causing build issues.

Fixes: #1839638, #1855497 and #1855491